added check before installing sops and a small fix

+++ homelab/bash/setup-trinity-helios/run-in-helios-setup-lxc.sh
@@ -74,10 +74,18 @@ function base_installation ()
     echo "------------------------------------------------------------------------"
     apt-get install -y curl age git ansible sshpass
 
-    SOPS_LATEST_VERSION=$(curl -s "https://api.github.com/repos/getsops/sops/releases/latest" | grep -Po '"tag_name": "v\K[0-9.]+')
-    curl -Lo sops.deb "https://github.com/getsops/sops/releases/download/v${SOPS_LATEST_VERSION}/sops_${SOPS_LATEST_VERSION}_amd64.deb"
-    dpkg -i ./sops.deb
-    apt-get install -f # Install missing dependencies
+    if ! dpkg -s sops &> /dev/null; then
+        echo "SOPS is not installed. Installing now..."
+        SOPS_LATEST_VERSION=$(curl -s "https://api.github.com/repos/getsops/sops/releases/latest" | grep -Po '"tag_name": "v\K[0-9.]+')
+        curl -Lo sops.deb "https://github.com/getsops/sops/releases/download/v${SOPS_LATEST_VERSION}/sops_${SOPS_LATEST_VERSION}_amd64.deb"
+        dpkg -i ./sops.deb
+        apt-get install -f # Install missing dependencies
+        echo "SOPS has been installed successfully."
+        sops --version
+    else
+        echo "SOPS is already installed."
+        sops --version
+    fi
 
     echo ""
     echo "--------------------------------------------------------"
@@ -89,8 +97,12 @@ function base_installation ()
     echo "--------------"
     echo "Cleaning up..."
     echo "--------------"
-    rm -rf sops.deb
-    echo "Removed installation files that are no longer required."
+    if [ -f sops.deb ]; then
+        rm -rf sops.deb
+        echo "Removed installation files that are no longer required."
+    else
+        echo "No installation files to remove."
+    fi
 }
 
 function github_pat_setup ()
@@ -236,7 +248,7 @@ function sops_decryption ()
 
     HELIOS_SETUP_ANSIBLE_DIR="$HOME/helios-setup/dotfiles-and-homelab/homelab/ansible"
     sops --decrypt --age $(cat $SOPS_AGE_KEY_FILE |grep -oP "public key: \K(.*)") \
-    -i "${HELIOS_SETUP_ANSIBLE_DIR}/inventory/group_vars/tailscale-auth/secrets.yml"
+    -i "${HELIOS_SETUP_ANSIBLE_DIR}/inventory/group_vars/tailscale_auth/secrets.yml"
 
     echo "Finished decrypting secrets with SOPS."
 }
